+++ game-blueprince/index.js
@@ -2,8 +2,8 @@
 Name: Blue Prince Vortex Extension
 Structure: Unity BepinEx/MelonLoader Hybrid
 Author: ChemBoy1
-Version: 0.1.0
-Date: 2025-10-31
+Version: 0.1.1
+Date: 2026-03-15
 //////////////////////////////////////////*/
 
 //Import libraries
@@ -11,6 +11,7 @@ const { actions, fs, util, selectors, log } = require('vortex-api');
 const path = require('path');
 const template = require('string-template');
 const { parseStringPromise } = require('xml2js');
+const winapi = require('winapi-bindings');
 
 const USER_HOME = util.getVortexPath("home");
 //const DOCUMENTS = util.getVortexPath("documents");
@@ -39,8 +40,8 @@ const XBOX_SAVE_STRING = '9s0pnehqffj7t'; //string after "ID_"
 const BEPINEX_BUILD = 'il2cpp'; // 'mono' or 'il2cpp'
 const ARCH = 'x64'; //'x64' or 'x86' game architecture (64-bit or 32-bit)
 const BEP_VER = '5.4.23.4'; //set BepInEx version for mono URLs
-const BEP_BE_VER = '738'; //set BepInEx build for BE URLs
-const BEP_BE_COMMIT = 'af0cba7'; //git commit number for BE builds
+const BEP_BE_VER = '755'; //set BepInEx build for BE URLs
+const BEP_BE_COMMIT = '3fab71a'; //git commit number for BE builds
 const allowBepCfgMan = false; //should BepInExConfigManager be downloaded?
 const allowMelPrefMan = false; //should MelonPreferencesManager be downloaded? False until figure out UniverseLib dependency
 const allowBepinexNexus = false; //set false until bugs are fixed
@@ -74,7 +75,7 @@ if (BEPINEX_BUILD === 'mono') {
 }
 const BEPINEX_FOLDER = 'BepInEx';
 const BEP_STRING = 'BepInEx';
-const BEP_PATCHER_STRING = 'BaseUnityPlugin'; // TODO: need to find reliable string
+const BEP_PATCHER_STRING = 'BepInEx.Preloader.Core.Patching';
 
 let BEPINEX_URL = `https://builds.bepinex.dev/projects/bepinex_be/${BEP_BE_VER}/BepInEx-Unity.IL2CPP-win-x64-6.0.0-be.${BEP_BE_VER}%2B${BEP_BE_COMMIT}.zip`;
 let BEPINEX_URL_ERR = `https://builds.bepinex.dev/projects/bepinex_be`;
@@ -98,6 +99,10 @@ const MELON_FILE = 'MelonLoader.dll';
 const MELON_FOLDER = 'MelonLoader';
 const MEL_STRING = 'MelonLoader';
 const MEL_PLUGIN_STRING = 'MelonPlugin';
+const MELON_DOTNET_VER = '6';
+const MELON_DOTNET_URL = `https://dotnet.microsoft.com/download/dotnet/${MELON_DOTNET_VER}.0`; //required for MelonLoader on IL2CPP games
+const DOTNET_REG_HIVE = 'HKEY_LOCAL_MACHINE';
+const DOTNET_REG_KEY = `SOFTWARE\\WOW6432Node\\dotnet\\Setup\\InstalledVersions\\x64\\sharedfx\\Microsoft.WindowsDesktop.App`;
 
 const ROOT_ID = `${GAME_ID}-root`;
 const ROOT_NAME = "Root Game Folder";
@@ -810,8 +815,7 @@ async function installPlugin(api, gameSpec, files, workingDir) {
         const content = await fs.readFileAsync(path.join(workingDir, file), 'utf8');
         if (content.includes(BEP_STRING)) {
             isBepinex = true;
-            isBepinexPatcher = false; //temporary, find reliable string to id patchers
-            //isBepinexPatcher = !content.includes(BEP_PATCHER_STRING) && !files.find(file => path.extname(file).toLowerCase() = BEPINEX_PLUGINS_FOLDER);
+            isBepinexPatcher = content.includes(BEP_PATCHER_STRING);
         } else if (content.includes(MEL_STRING)) {
           isMelon = true;
           isMelonPlugin = content.includes(MEL_PLUGIN_STRING);
@@ -963,8 +967,8 @@ async function chooseModLoader(api, gameSpec) {
       { replace }
     ),
   }, [
+    { label: t('MelonLoader (Recommended)') },
     { label: t('BepInEx') },
-    { label: t('MelonLoader') },
   ])
   .then(async (result) => {
     if (result === undefined) {
@@ -972,7 +976,7 @@ async function chooseModLoader(api, gameSpec) {
     }
     if (result.action === 'BepInEx') {
       await downloadBepinex(api, gameSpec);
-    } else if (result.action === 'MelonLoader') {
+    } else if (result.action === 'MelonLoader (Recommended)') {
       await downloadMelon(api, gameSpec);
     }
   });
@@ -1162,6 +1166,77 @@ async function modFoldersEnsureWritable(gamePath, relPaths) {
   }
 }
 
+function dotNetMelonNotify(api) {
+  const NOTIF_ID = `${GAME_ID}-dotnetmelon-notify`;
+  const MESSAGE = `.NET ${MELON_DOTNET_VER} Required`;
+  api.sendNotification({
+    id: NOTIF_ID,
+    type: 'warning',
+    message: MESSAGE,
+    allowSuppress: true,
+    actions: [
+      {
+        title: `Download .NET ${MELON_DOTNET_VER}`,
+        action: (dismiss) => {
+          util.opn(MELON_DOTNET_URL).catch(() => null);
+          dismiss();
+        }
+      },
+      {
+        title: 'More',
+        action: (dismiss) => {
+          api.showDialog('question', MESSAGE, {
+            text: `\n`
+                + `MelonLoader requires .NET ${MELON_DOTNET_VER} to be installed on your system for IL2CPP build Unity games, like this game.\n`
+                + `\n`
+                + `Please install .NET ${MELON_DOTNET_VER} so that MelonLoader can function. Your game may crash at launch if the correct version of .NET is not installed.\n`
+                + `\n`
+          }, [
+            { label: `Download .NET ${MELON_DOTNET_VER}`, action: () => {
+              util.opn(MELON_DOTNET_URL).catch(() => null);
+              dismiss();
+            }},
+            { label: 'Not Now', action: () => dismiss() },
+            {
+              label: 'Never Show Again', action: () => {
+                api.suppressNotification(NOTIF_ID);
+                dismiss();
+              }
+            },
+          ]);
+        },
+      },
+    ],
+  });
+}
+
+async function checkDotNetMelon(api) {
+  const version = MELON_DOTNET_VER;
+  let values = undefined;
+  try {
+    const buffer = winapi.WithRegOpen( //array of objects with values.type and values.key
+      DOTNET_REG_HIVE,
+      DOTNET_REG_KEY,
+      (hkey) => { //have to enum in the callback - https://github.com/Nexus-Mods/node-winapi-bindings/blob/master/index.d.ts
+        values = winapi.RegEnumValues(hkey); //array of objects with values.type and values.key
+      }
+    );
+    if (!values) {
+      dotNetMelonNotify(api); //assume not installed if key not found
+    }
+    values = values.map(value => value.key); //map array to only keys
+    const found = values.some(value => value.startsWith(version)); //find entry starting with correct version number
+    if (found) {
+      //log('warn', `Found .NET ${version} installation`);
+    } else {
+      dotNetMelonNotify(api); //assume not installed if key not found
+    }
+  } catch (err) { //*/
+    log('warn', `Failed to read .NET registry key: ${err}`);
+    dotNetMelonNotify(api)
+  }
+}
+
 //Setup function
 async function setup(discovery, api, gameSpec) {
   //SYNC CODE ////////////////////////////////////
@@ -1183,6 +1258,9 @@ async function setup(discovery, api, gameSpec) {
   } //*/
   if (melonInstalled && allowMelPrefMan) {
     await downloadMelonPrefManNotify(api, gameSpec); //notification to download MelonPreferencesManager
+  } //*/
+  if (isMelonInstalled(api, gameSpec)) {
+    checkDotNetMelon(api); //check for .NET 6 installation
   } //*/
   return modFoldersEnsureWritable(GAME_PATH, MODTYPE_FOLDERS);
 }
@@ -1330,6 +1408,9 @@ function main(context) {
       } //*/
       if (melonInstalled && allowMelPrefMan) {
         await downloadMelonPrefMan(context.api, spec); //download MelonPreferencesManager
+      } //*/
+      if (isMelonInstalled(context.api, spec)) {
+        checkDotNetMelon(context.api); //check for .NET 6 installation
       } //*/
     });
   });