+++ game-megabonk/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - None
 
+## [0.1.2] - 2026-03-15
+
+- Fixed: Bump BepInEx BE version to 755
+- Added: .NET 6 installation check and notification for MelonLoader
+- Added: Automatic detection of BepInEx patchers
+
 ## [0.1.1] - 2025-11-20
 
 - Added support for CustomCharacters mods. Install folder is set based on which mod loader is installed.